+++ src/View.jsx
@@ -43,7 +43,7 @@ class View extends Component {
     const { viewsToShow, axis } = pager.options
     const { tag, trackSize, ...restProps } = this.props
     let style = {
-      ...this.props.style
+      ...restProps.style
     }
 
     // hide view visually until it has been added to the pager
@@ -53,28 +53,10 @@ class View extends Component {
       style.pointerEvents = 'none'
     }
 
-    // we need to position views inline when using the x axis
-    if (axis === 'x') {
-      style.display = 'inline-block'
-      style.verticalAlign = 'top'
-    }
-
-    // set width or height on view when viewsToShow is not auto
-    if (viewsToShow !== 'auto' && pager.views.length) {
-      style[axis === 'x' ? 'width' : 'height'] = 100 / pager.views.length + '%'
-    }
-
     if (this._viewInstance) {
-      // make sure view stays in frame when using infinite option
-      if (pager.options.infinite && !this._viewInstance.inBounds) {
-        style.position = 'relative'
-        style[(axis === 'y') ? 'top' : 'left'] = this._viewInstance.getPosition()
-      }
-
-      // apply any animations
       style = {
         ...style,
-        ...pager.animationBus.getStyles(this._viewInstance)
+        ...this._viewInstance.getStyles()
       }
     }