+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### 0.5.0-prerelease.4
 Use [resize-observer-polyfill](https://github.com/que-etc/resize-observer-polyfill) to make sure the pager is always calculated with proper view dimensions
 
+Hide view visually until it has been added to the pager
+
 ### 0.5.0-prerelease.3
 emit `viewChange` when `viewsToShow` options have changed