+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## CHANGELOG
+### 0.5.0-prerelease.5
+Fix old `destroy` method left over from window resize
+
 ### 0.5.0-prerelease.4
 Use [resize-observer-polyfill](https://github.com/que-etc/resize-observer-polyfill) to make sure the pager is always calculated with proper view dimensions